+++ src/pixelToWorld.cpp
@@ -29,7 +29,7 @@ std::vector<double> PixelToWorld:: worldPoints(MatrixXf T, std::vector<double> p
 std::vector<double> real_world;
 MatrixXf pseudo_T(4,3);
 pseudo_T= T.completeOrthogonalDecomposition().pseudoInverse();
-for (int i = 0; i<pixelValues.size(); i=i+2) {
+for (long unsigned int i = 0; i<pixelValues.size(); i=i+2) {
 double u = pixelValues.at(i);
 double v = pixelValues.at(i+1);
 Vector3f pixel_coord;
